+++ src/polymer/polymer-element-mixin.ts
@@ -67,7 +67,7 @@ export class ScannedPolymerElementMixin extends ScannedElementMixin implements
     Object.assign(element, this);
 
     for (const method of element.methods) {
-      // methods are only public by default if they're documented.
+      // Method privacy is infered from function names unless specifically defined in jsdoc
       method.privacy = getOrInferPrivacy(method.name, method.jsdoc, true);
     }
     element.mixins = [];